登録できる

+++ lib/model/db/logDB.dart
@@ -16,7 +16,7 @@ final logProvider = Provider<Future<Database>>((ref) async {
       // 初回作成時のテーブル作成処理
       await db.execute('''
         CREATE TABLE items(
-          dvancedDist INT,
+          dvancedDist DOUBLE,
           createdAt TEXT                          
         )
       ''');
